refactor: import code accordingly to previous commit
