+++ hrm-domain/hrm-service/src/permissions/canPerformActionOnObject.ts
@@ -16,7 +16,7 @@
 
 import { TwilioUser } from '@tech-matters/twilio-worker-auth';
 import { Actions, TargetKind, isValidSetOfActionsForTarget } from './actions';
-import { InitializedCan } from './setupCanForRules';
+import { InitializedCan } from './initializeCanForRules';
 import { getContactById } from '../contact/contactService';
 import { getCase as getCaseById } from '../case/caseService';
 import { assertExhaustive } from '../contact-job/assertExhaustive';
